+++ appwrite/enums/backup_services.py
@@ -2,5 +2,8 @@ from enum import Enum
 
 class BackupServices(Enum):
     DATABASES = "databases"
+    TABLESDB = "tablesdb"
+    DOCUMENTSDB = "documentsdb"
+    VECTORSDB = "vectorsdb"
     FUNCTIONS = "functions"
     STORAGE = "storage"