+++ appwrite/enums/tables_db_index_type.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-class IndexType(Enum):
+class TablesDBIndexType(Enum):
     KEY = "key"
     FULLTEXT = "fulltext"
     UNIQUE = "unique"