+++ appwrite/models/__init__.py
@@ -21,6 +21,7 @@ from .framework_list import FrameworkList
 from .runtime_list import RuntimeList
 from .deployment_list import DeploymentList
 from .execution_list import ExecutionList
+from .webhook_list import WebhookList
 from .country_list import CountryList
 from .continent_list import ContinentList
 from .language_list import LanguageList
@@ -107,6 +108,7 @@ from .framework import Framework
 from .framework_adapter import FrameworkAdapter
 from .deployment import Deployment
 from .execution import Execution
+from .webhook import Webhook
 from .variable import Variable
 from .country import Country
 from .continent import Continent
@@ -163,6 +165,7 @@ __all__ = [
     'RuntimeList',
     'DeploymentList',
     'ExecutionList',
+    'WebhookList',
     'CountryList',
     'ContinentList',
     'LanguageList',
@@ -249,6 +252,7 @@ __all__ = [
     'FrameworkAdapter',
     'Deployment',
     'Execution',
+    'Webhook',
     'Variable',
     'Country',
     'Continent',